refactor: moved models and validation out of schema
